refactor: make a hierarchy on X3dRepository methods before. Data level operations are now handled in X3dDataSource
